Add advent calendar view for ES6/2015 calendar
