+++ content/en/events/2015/advent-calendar/index.md
@@ -1,0 +1,243 @@
+---
+layout: AdventCalendar
+date: "2015-12-01"
+title: "Advent calendar 2015: ES6/ES2015 in 24 days!"
+tags:
+  - JavaScript
+  - ES6
+  - ES2015
+header:
+  image: https://farm4.staticflickr.com/3215/3132286400_33fdcab077_z.jpg
+  credit: https://www.flickr.com/photos/dvids/7112879347
+  linearGradient: 160deg, rgba(204, 51, 51, .8), rgba(204, 51, 51, .4)
+---
+
+We started a crazy challenge for a December 1 (yes we decided on the 1st
+at 3pm): **teach you ES6 (ES2015) in 24 days**.
+
+We'll be a little on the ball for a few days (and articles might not be
+published the morning), but the result should be nice!
+
+<div class="r-Grid">
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href="/en/articles/js/es2015/const-let-var/"
+    >
+      1. Declaring variables with const, let and var
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      2. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      3. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      4. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      5. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      6. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      7. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      8. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      9. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      10. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      11. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      12. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      13. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      14. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      15. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      16. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      17. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      18. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      19. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      20. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      21. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      22. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      23. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      24. ?
+    </a>
+  </div>
+</div>
+
+<p style="font-size: 0.8rem; margin: 2rem; text-align: center;">
+  <a href="https://github.com/putaindecode/putaindecode.io/issues/553">
+    If you want to help, it's over here.
+  </a>
+</p>
