+++ content/fr/evenements/2015/calendrier-avent/index.md
@@ -1,0 +1,243 @@
+---
+layout: AdventCalendar
+date: "2015-12-01"
+title: "Calendrier de l'Avent 2015: ES6/ES2015 en 24 jours !"
+tags:
+  - JavaScript
+  - ES6
+  - ES2015
+header:
+  image: https://farm4.staticflickr.com/3215/3132286400_33fdcab077_z.jpg
+  credit: https://www.flickr.com/photos/dvids/7112879347
+  linearGradient: 160deg, rgba(204, 51, 51, .8), rgba(204, 51, 51, .4)
+---
+
+On s'est lancés un défi un peu fou pour un 1er décembre (oui on a décidé le 1er
+à 15h) : **vous apprendre ES6 (ES2015) en 24 jours.**
+
+On va être un peu sur les rotules quelques jours (et les articles sortiront pas
+forcément le matin), mais le résultat devrait être sympathique !
+
+<div class="r-Grid">
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href="/fr/articles/js/es2015/const-let-var/"
+    >
+      1. La déclaration de variables avec const, let et var
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href="/fr/articles/js/es2015/destructuring/"
+    >
+      2. Le destructuring
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      3. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      4. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      5. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      6. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      7. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      8. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      9. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      10. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      11. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      12. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      13. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      14. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      15. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      16. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      17. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      18. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      19. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      20. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      21. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      22. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      23. ?
+    </a>
+  </div>
+
+  <div class="r-Grid-cell r-minM--1of4">
+    <a
+      class="putainde-Day"
+      href=""
+    >
+      24. ?
+    </a>
+  </div>
+</div>
+
+<p style="font-size: 0.8rem; margin: 2rem; text-align: center;">
+  <a href="https://github.com/putaindecode/putaindecode.io/issues/553">
+    Si vous voulez aidez, c'est par ici
+  </a>
+</p>
