+++ web_modules/_legacy-css/index.css
@@ -50,3 +50,30 @@ html {
 /* @import "./js/images.css"; */
 
 @import "./blocks/highlight.css";
+
+/* advent calendar */
+.putainde-Day {
+  padding: .5rem;
+  border: 1px solid #c33;
+  border-radius: 2px;
+  display: block;
+  min-height: 7rem;
+  margin: .2rem;
+}
+
+.putainde-Day[href=""] {
+  opacity: .4;
+}
+
+.putainde-AdventCalendarLink {
+  display: block;
+  font-size: 2.6rem;
+  line-height: 6rem;
+  margin: 2rem;
+  padding: 0 0 3rem;
+
+  border-bottom: 1px solid gray;
+  color: #c33;
+  text-align: center;
+  text-decoration: none;
+}
