+++ web_modules/app/pageComponents.js
@@ -3,3 +3,4 @@ export PageError from "Layout/PageError"
 export Homepage from "Layout/Homepage"
 export Post from "Layout/Post"
 export Posts from "Layout/Posts"
+export AdventCalendar from "Layout/AdventCalendar"
